+++ selfdrive/test/process_replay/regen.py
@@ -20,6 +20,7 @@ from selfdrive.car.toyota.values import EPS_SCALE
 from selfdrive.manager.process import ensure_running
 from selfdrive.manager.process_config import managed_processes
 from selfdrive.test.process_replay.process_replay import CONFIGS, FAKEDATA, setup_env, check_openpilot_enabled
+from selfdrive.test.process_replay.migration import migrate_all
 from selfdrive.test.update_ci_routes import upload_route
 from tools.lib.route import Route
 from tools.lib.framereader import FrameReader
@@ -179,67 +180,11 @@ def replay_cameras(lr, frs, disable_tqdm=False):
   return vs, p
 
 
-def migrate_carparams(lr):
-  all_msgs = []
-  for msg in lr:
-    if msg.which() == 'carParams':
-      CP = messaging.new_message('carParams')
-      CP.carParams = msg.carParams.as_builder()
-      for car_fw in CP.carParams.carFw:
-        car_fw.brand = CP.carParams.carName
-      msg = CP.as_reader()
-    all_msgs.append(msg)
-
-  return all_msgs
-
-
-def migrate_sensorEvents(lr, old_logtime=False):
-  all_msgs = []
-  for msg in lr:
-    if msg.which() != 'sensorEventsDEPRECATED':
-      all_msgs.append(msg)
-      continue
-
-    # migrate to split sensor events
-    for evt in msg.sensorEventsDEPRECATED:
-      # build new message for each sensor type
-      sensor_service = ''
-      if evt.which() == 'acceleration':
-        sensor_service = 'accelerometer'
-      elif evt.which() == 'gyro' or evt.which() == 'gyroUncalibrated':
-        sensor_service = 'gyroscope'
-      elif evt.which() == 'light' or evt.which() == 'proximity':
-        sensor_service = 'lightSensor'
-      elif evt.which() == 'magnetic' or evt.which() == 'magneticUncalibrated':
-        sensor_service = 'magnetometer'
-      elif evt.which() == 'temperature':
-        sensor_service = 'temperatureSensor'
-
-      m = messaging.new_message(sensor_service)
-      m.valid = True
-      if old_logtime:
-        m.logMonoTime = msg.logMonoTime
-
-      m_dat = getattr(m, sensor_service)
-      m_dat.version = evt.version
-      m_dat.sensor = evt.sensor
-      m_dat.type = evt.type
-      m_dat.source = evt.source
-      if old_logtime:
-        m_dat.timestamp = evt.timestamp
-      setattr(m_dat, evt.which(), getattr(evt, evt.which()))
-
-      all_msgs.append(m.as_reader())
-
-  return all_msgs
-
-
 def regen_segment(lr, frs=None, daemons="all", outdir=FAKEDATA, disable_tqdm=False):
   if not isinstance(daemons, str) and not hasattr(daemons, "__iter__"):
     raise ValueError("whitelist_proc must be a string or iterable")
 
-  lr = migrate_carparams(list(lr))
-  lr = migrate_sensorEvents(list(lr))
+  lr = migrate_all(lr)
   if frs is None:
     frs = dict()
